Move Memory.Array to Memory.Mutable.Array
