+++ README.md
@@ -368,7 +368,7 @@ documentation.
 
 ## Arrays
 
-The `Streamly.Memory.Array` module provides immutable arrays.  Arrays are the
+The `Streamly.Memory.Mutable.Array` module provides immutable arrays.  Arrays are the
 computing duals of streams. Streams are good at sequential access and immutable
 transformations of in-transit data whereas arrays are good at random access and
 in-place transformations of buffered data. Unlike streams which are potentially
@@ -379,7 +379,7 @@ computing system. The design of streamly as a general purpose computing
 framework is centered around these two fundamental aspects of computing and
 storage.
 
-`Streamly.Memory.Array` uses pinned memory outside GC and therefore avoid any
+`Streamly.Memory.Mutable.Array` uses pinned memory outside GC and therefore avoid any
 GC overhead for the storage in arrays. Streamly allows efficient
 transformations over arrays using streams. It uses arrays to transfer data to
 and from the operating system and to store data in memory.
@@ -422,7 +422,7 @@ module Main where
 
 import qualified Streamly.Prelude as S
 import qualified Streamly.Data.Fold as FL
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Memory.Mutable.Array as A
 import qualified Streamly.FileSystem.Handle as FH
 import qualified System.IO as FH
