+++ benchmark/Array.hs
@@ -13,8 +13,8 @@ import System.Random (randomRIO)
 import qualified GHC.Exts as GHC
 
 import qualified ArrayOps as Ops
-import qualified Streamly.Internal.Memory.Array as IA
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Internal.Memory.Mutable.Array as IA
+import qualified Streamly.Memory.Mutable.Array as A
 import qualified Streamly.Prelude as S
 
 import Gauge
