+++ benchmark/Streamly/Benchmark/Data/Prim/Array.hs
@@ -81,7 +81,7 @@ main =
 #if 0
         -- PrimArray does not have a Foldable instance because it requires a
         -- Prim constraint. Though it should be possible to make an instance in
-        -- the same way as we do in Memory.Array.
+        -- the same way as we do in Memory.Mutable.Array.
         , benchPureSink "foldable/foldl'" Ops.foldableFoldl'
         , benchPureSink "foldable/sum" Ops.foldableSum
 #endif
