+++ src/Streamly/Data/Unicode/Stream.hs
@@ -13,11 +13,11 @@
 -- A 'Char' stream is the canonical representation to process Unicode strings.
 -- It can be processed efficiently using regular stream processing operations.
 -- A byte stream of Unicode text read from an IO device or from an
--- 'Streamly.Memory.Array.Array' in memory can be decoded into a 'Char' stream
+-- 'Streamly.Memory.Mutable.Array.Array' in memory can be decoded into a 'Char' stream
 -- using the decoding routines in this module.  A 'String' (@[Char]@) can be
 -- converted into a 'Char' stream using 'Streamly.Prelude.fromList'.  An @Array
 -- Char@ can be 'Streamly.Prelude.unfold'ed into a stream using the array
--- 'Streamly.Memory.Array.read' unfold.
+-- 'Streamly.Memory.Mutable.Array.read' unfold.
 --
 -- = Storing Unicode Strings
 --
@@ -31,7 +31,7 @@
 -- than pinned arrays for short and short lived strings.
 --
 -- For longer or long lived streams you can 'Streamly.Prelude.fold' the 'Char'
--- stream as @Array Char@ using the array 'Streamly.Memory.Array.write' fold.
+-- stream as @Array Char@ using the array 'Streamly.Memory.Mutable.Array.write' fold.
 -- The 'Array' type provides a more compact representation and pinned memory
 -- reducing GC overhead. If space efficiency is a concern you can use
 -- 'encodeUtf8' on the 'Char' stream before writing it to an 'Array' providing
