+++ src/Streamly/Internal/Data/SmallArray.hs
@@ -103,7 +103,7 @@ foldr f z arr = runIdentity $ D.foldr f z $ toStreamD arr
 -- 'SmallArray'.
 --
 -- Since we are folding to a 'SmallArray' @n@ should be <= 128, for larger number
--- of elements use an 'Array' from either "Streamly.Data.Array" or "Streamly.Memory.Array".
+-- of elements use an 'Array' from either "Streamly.Data.Array" or "Streamly.Memory.Mutable.Array".
 {-# INLINE_NORMAL writeN #-}
 writeN :: MonadIO m => Int -> Fold m a (SmallArray a)
 writeN limit = Fold step initial extract
@@ -135,7 +135,7 @@ fromStreamDN limit str = do
 --
 -- It is recommended to use a value of @n@ <= 128. For larger sized
 -- arrays, use an 'Array' from "Streamly.Data.Array" or
--- "Streamly.Memory.Array"
+-- "Streamly.Memory.Mutable.Array"
 {-# INLINABLE fromListN #-}
 fromListN :: Int -> [a] -> SmallArray a
 fromListN n xs = unsafePerformIO $ fromStreamDN n $ D.fromList xs
