+++ src/Streamly/Internal/FileSystem/File.hs
@@ -104,7 +104,7 @@ import qualified System.IO as SIO
 
 import Streamly.Internal.Data.Fold.Types (Fold(..))
 import Streamly.Internal.Data.Unfold.Types (Unfold(..))
-import Streamly.Internal.Memory.Array.Types
+import Streamly.Internal.Memory.Mutable.Array.Types
        (Array(..), defaultChunkSize, writeNUnsafe)
 import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Stream.StreamK.Type (IsStream)
@@ -115,8 +115,8 @@ import Streamly.Internal.Data.SVar (MonadAsync)
 import qualified Streamly.Internal.Data.Fold.Types as FL
 import qualified Streamly.Internal.Data.Unfold as UF
 import qualified Streamly.Internal.FileSystem.Handle as FH
-import qualified Streamly.Internal.Memory.ArrayStream as AS
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Internal.Memory.Mutable.ArrayStream as AS
+import qualified Streamly.Memory.Mutable.Array as A
 import qualified Streamly.Prelude as S
 
 -------------------------------------------------------------------------------
@@ -236,7 +236,7 @@ readWithBufferOf = UF.concat (usingFilexxx FH.readChunksWithBufferOf) A.read
 
 -- | Unfolds a file path into a byte stream. IO requests to the device are
 -- performed in sizes of
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize'.
+-- 'Streamly.Internal.Memory.Mutable.Array.Types.defaultChunkSize'.
 --
 -- @since 0.7.0
 {-# INLINE read #-}
@@ -376,7 +376,7 @@ writeWithBufferOf n path =
 -- > write = 'writeWithBufferOf' A.defaultChunkSize
 --
 -- | Write a byte stream to a file. Accumulates the input in chunks of up to
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize' before writing to
+-- 'Streamly.Internal.Memory.Mutable.Array.Types.defaultChunkSize' before writing to
 -- the IO device.
 --
 -- /Internal/
