+++ src/Streamly/Internal/Memory/Mutable/Array/Types.hs
@@ -10,7 +10,7 @@
 #include "inline.hs"
 
 -- |
--- Module      : Streamly.Internal.Memory.Array.Types
+-- Module      : Streamly.Internal.Memory.Mutable.Array.Types
 -- Copyright   : (c) 2019 Composewell Technologies
 --
 -- License     : BSD3
@@ -18,7 +18,7 @@
 -- Stability   : experimental
 -- Portability : GHC
 --
-module Streamly.Internal.Memory.Array.Types
+module Streamly.Internal.Memory.Mutable.Array.Types
     (
       Array (..)
 
@@ -684,7 +684,7 @@ fromStreamDArraysOf n (D.Stream step state) =
     step' _ (GroupStart st) = do
         when (n <= 0) $
             -- XXX we can pass the module string from the higher level API
-            error $ "Streamly.Internal.Memory.Array.Types.fromStreamDArraysOf: the size of "
+            error $ "Streamly.Internal.Memory.Mutable.Array.Types.fromStreamDArraysOf: the size of "
                  ++ "arrays [" ++ show n ++ "] must be a natural number"
         Array start end bound <- liftIO $ newArray n
         return $ D.Skip (GroupBuffer st start end bound)
@@ -1099,7 +1099,7 @@ packArraysChunksOf n (D.Stream step state) =
     step' gst (SpliceInitial st) = do
         when (n <= 0) $
             -- XXX we can pass the module string from the higher level API
-            error $ "Streamly.Internal.Memory.Array.Types.packArraysChunksOf: the size of "
+            error $ "Streamly.Internal.Memory.Mutable.Array.Types.packArraysChunksOf: the size of "
                  ++ "arrays [" ++ show n ++ "] must be a natural number"
         r <- step gst st
         case r of
@@ -1145,7 +1145,7 @@ lpackArraysChunksOf n (Fold step1 initial1 extract1) =
     initial = do
         when (n <= 0) $
             -- XXX we can pass the module string from the higher level API
-            error $ "Streamly.Internal.Memory.Array.Types.packArraysChunksOf: the size of "
+            error $ "Streamly.Internal.Memory.Mutable.Array.Types.packArraysChunksOf: the size of "
                  ++ "arrays [" ++ show n ++ "] must be a natural number"
         r1 <- initial1
         return (Tuple' Nothing r1)
@@ -1204,11 +1204,11 @@ groupIOVecsOf n maxIOVLen (D.Stream step state) =
     step' gst (GatherInitial st) = do
         when (n <= 0) $
             -- XXX we can pass the module string from the higher level API
-            error $ "Streamly.Internal.Memory.Array.Types.groupIOVecsOf: the size of "
+            error $ "Streamly.Internal.Memory.Mutable.Array.Types.groupIOVecsOf: the size of "
                  ++ "groups [" ++ show n ++ "] must be a natural number"
         when (maxIOVLen <= 0) $
             -- XXX we can pass the module string from the higher level API
-            error $ "Streamly.Internal.Memory.Array.Types.groupIOVecsOf: the number of "
+            error $ "Streamly.Internal.Memory.Mutable.Array.Types.groupIOVecsOf: the number of "
                  ++ "IOVec entries [" ++ show n ++ "] must be a natural number"
         r <- step (adaptState gst) st
         case r of
