+++ src/Streamly/Internal/Memory/Unicode/Array.hs
@@ -22,10 +22,10 @@ where
 import Control.Monad.IO.Class (MonadIO)
 import Streamly (IsStream, MonadAsync)
 import Prelude hiding (String, lines, words, unlines, unwords)
-import Streamly.Memory.Array (Array)
+import Streamly.Memory.Mutable.Array (Array)
 
 import qualified Streamly.Internal.Data.Unicode.Stream as S
-import qualified Streamly.Memory.Array as A
+import qualified Streamly.Memory.Mutable.Array as A
 
 -- | Break a string up into a stream of strings at newline characters.
 -- The resulting strings do not contain newlines.
