+++ src/Streamly/Network/Socket.hs
@@ -18,9 +18,9 @@
 -- = Programmer Notes
 --
 -- Read IO requests to connected stream sockets are performed in chunks of
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize'.  Unless specified
+-- 'Streamly.Internal.Memory.Mutable.Array.Types.defaultChunkSize'.  Unless specified
 -- otherwise in the API, writes are collected into chunks of
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize' before they are
+-- 'Streamly.Internal.Memory.Mutable.Array.Types.defaultChunkSize' before they are
 -- written to the socket. APIs are provided to control the chunking behavior.
 --
 -- > import qualified Streamly.Network.Socket as SK
@@ -29,7 +29,7 @@
 -- "Streamly.Internal.Network.Socket" module.
 
 -- By design, connected socket IO APIs are similar to
--- "Streamly.Memory.Array" read write APIs. They are almost identical to the
+-- "Streamly.Memory.Mutable.Array" read write APIs. They are almost identical to the
 -- sequential streaming APIs in "Streamly.Internal.FileSystem.File".
 --
 module Streamly.Network.Socket
